move: data package -> data module
