+++ data/src/main/java/com/oreocube/booksearch/data/di/RepositoryModule.kt
@@ -1,4 +1,4 @@
-package com.oreocube.booksearch.di
+package com.oreocube.booksearch.data.di
 
 import com.oreocube.booksearch.data.repository.FavoriteRepositoryImpl
 import com.oreocube.booksearch.data.repository.LibraryRepositoryImpl
