+++ data/src/main/java/com/oreocube/booksearch/data/response/BookAvailabilityResponse.kt
