+++ data/src/main/java/com/oreocube/booksearch/data/response/BookDetailResponse.kt
