+++ data/src/main/java/com/oreocube/booksearch/data/response/LibraryShortEntity.kt
